Updated user post and put functions to consolodate user validation
