+++ uhc.js
@@ -22,7 +22,8 @@
     repositories = require('./repository/repository'),
     exception = require('./exception'),
     security = require('./security'),
-    model = require('./model/model');
+    model = require('./model/model'),
+    User = require('./model/User');
 
  const repository = new repositories.UhcRepositories(config.db.server);
  /**
@@ -140,6 +141,7 @@
             throw new exception.Exception("Error refreshing session", exception.ErrorCodes.SECURITY_ERROR, e);
         }
     }
+
     /**
      * @method
      * @summary Register a regular user 
@@ -150,6 +152,11 @@
 
         // First we register the user in our DB
         try {
+
+            // Validate the user
+            this.validateUser(user, password);
+
+            // Insert the user
             var retVal = await repository.userRepository.insert(user, password);
 
             // TODO: Add user to group USERS
@@ -173,6 +180,50 @@
      */
     async registerExternalUser(identity) {
         throw new exception.NotImplementedException();
+    }
+
+    /**
+     * @method
+     * @summary Updates the specified user setting their new password if necessary
+     * @param {User} user The user to be updated
+     * @param {string} newPassword The new password to set the user account to
+     * @returns {User} The updated user
+     */
+    async updateUser(user, newPassword) {
+        try {
+
+            // Validate the user
+            this.validateUser(user, newPassword);
+
+            // Update the user
+            return await repository.userRepository.update(user, newPassword);
+        }
+        catch(e) {
+            console.error("Error updating user: " + e.message);
+            throw new exception.Exception("Error updating user", exception.ErrorCodes.UNKNOWN, e);
+        }
+    }
+
+    /**
+     * @method
+     * @summary Ensures that the user object is valid
+     * @param {User} user The user to be validated
+     * @param {string} password The password the user is attempting to set
+     * @returns {boolean} True if the user passed validation
+     * @throws {BusinessRuleViolationException} When there are problems with the user
+     */
+    validateUser(user, password) {
+
+        var ruleViolations = [];
+        if(user.name && !new RegExp(config.security.username_regex).test(user.name))
+            ruleViolations.push(new exception.RuleViolation("Username format invalid", exception.ErrorCodes.INVALID_USERNAME, exception.RuleViolationSeverity.ERROR));
+        if(password && !new RegExp(config.security.password_regex).test(password))
+            ruleViolations.push(new exception.RuleViolation("Password does not meet complexity requirements", exception.ErrorCodes.PASSWORD_COMPLEXITY, exception.RuleViolationSeverity.ERROR));
+
+        if(ruleViolations.length > 0)
+            throw new exception.BusinessRuleViolationException(ruleViolations);
+        
+        return true;
     }
  }
